// src/common/enums/trade-event-type.enum.ts
export enum TradeEventType {
  CANDIDATE_CREATED = 'CANDIDATE_CREATED',
  CANDIDATE_QUALIFIED = 'CANDIDATE_QUALIFIED',
  CANDIDATE_WAIT = 'CANDIDATE_WAIT',
  CANDIDATE_REJECTED = 'CANDIDATE_REJECTED',
  CANDIDATE_NOTIFIED = 'CANDIDATE_NOTIFIED',
  CANDIDATE_ANALYSIS_FAILED = 'CANDIDATE_ANALYSIS_FAILED',
  CANDIDATE_SKIPPED = 'CANDIDATE_SKIPPED',
  TRADE_OPENED = 'TRADE_OPENED',
  STOP_CHANGED = 'STOP_CHANGED',
  TRADE_CLOSED = 'TRADE_CLOSED',
  USER_NOTE = 'USER_NOTE',
  PRICE_MILESTONE_REACHED = 'PRICE_MILESTONE_REACHED',
  ADVERSE_MOVE_OBSERVED = 'ADVERSE_MOVE_OBSERVED',
  STOP_PROXIMITY_OBSERVED = 'STOP_PROXIMITY_OBSERVED',
  STOP_BREACHED_OBSERVED = 'STOP_BREACHED_OBSERVED',
  TARGET1_REACHED = 'TARGET1_REACHED',
  TARGET2_REACHED = 'TARGET2_REACHED',
}